pyproject.toml
[tool.poetry]
name = "rock-paper-scissor"
version = "0.1.0"
description = ""
authors = ["Alexandru Leorda <alexandru.leorda@gmail.com>"]
readme = "README.md"
packages = [{include = "rock_paper_scissor"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-django = "^4.5.2"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
